type adjustments

+++ cloud/albuns/src/main/java/br/com/ada/albuns/controller/AlbumTemplateController.java
@@ -3,6 +3,7 @@ package br.com.ada.albuns.controller;
 import java.util.List;
 
 import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -10,7 +11,6 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import org.springframework.http.MediaType;
 
 import br.com.ada.albuns.model.dto.AlbumTemplateDTO;
 import br.com.ada.albuns.service.AlbumTemplateService;
@@ -39,7 +39,7 @@ public class AlbumTemplateController {
   }
 
   @GetMapping("/{id}")
-  public ResponseEntity<AlbumTemplateDTO> findById(@PathVariable("id") int id) {
+  public ResponseEntity<AlbumTemplateDTO> findById(@PathVariable("id") Long id) {
       try {
           return ResponseEntity.ok(service.findById(id));
       } catch (EntityNotFoundException ex) {
