+++ cloud/albuns/src/main/java/br/com/ada/albuns/repository/AlbumTemplateRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import br.com.ada.albuns.model.entity.AlbumTemplate;
 
-public interface AlbumTemplateRepository extends JpaRepository<AlbumTemplate, Integer>{
+public interface AlbumTemplateRepository extends JpaRepository<AlbumTemplate, Long>{
   
 }
